Switch to workspace, use tar&zstd instead of zip
