+++ Cargo.toml
@@ -1,25 +1,14 @@
-[package]
-name = "proton-launch"
+
+[workspace]
+members = ["proton-launch-cli", "proton-launch"]
+
+default-members = ["proton-launch-cli"]
+
+[workspace.dependencies]
+clap = { version = "4.1", features = ["derive", "string"] }
+
+[workspace.package]
 version = "0.1.1"
 edition = "2021"
 description = "Launch Windows games with Proton"
 license = "MIT"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-clap = { version = "4.1", features = ["derive", "string"] }
-dirs = "4.0.0"
-exe = "0.5.6"
-image = "0.24.5"
-keyvalues-parser = "0.1.0"
-open = "4.0.0"
-serde = { version = "1.0.156", features = ["derive"] }
-serde_ini = "0.2.0"
-thiserror = "1.0.39"
-time = "0.3.20"
-walkdir = "2.3.3"
-xdg = "2.4"
-zip = "0.6.4"
-
-[workspace]
