+++ proton-launch/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "proton-launch"
+version.workspace = true
+description.workspace = true
+edition.workspace = true
+license.workspace = true
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+clap = { workspace = true, optional = true }
+dirs = "4.0.0"
+exe = "0.5.6"
+image = "0.24.5"
+keyvalues-parser = "0.1.0"
+open = "4.0.0"
+serde = { version = "1.0.156", features = ["derive"] }
+serde_ini = "0.2.0"
+thiserror = "1.0.39"
+time = "0.3.20"
+walkdir = "2.3.3"
+xdg = "2.4"
+zip = "0.6.4"
+toml = "0.7.3"
+zstd = "0.12.3"
+tar = "0.4.38"
+
+[features]
+commandline = ["clap"]
