+++ proton-launch/src/lib.rs
@@ -2,3 +2,4 @@ pub mod command;
 pub mod paths;
 pub mod proton;
 pub mod steam;
+pub mod shortcut;
